% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_wiki.R
\name{get_section_number}
\alias{get_section_number}
\title{Get the Section Number for a Wiktionary Page}
\usage{
get_section_number(word, section = "Etymology")
}
\arguments{
\item{word}{Character; the Wiktionary page in question.}

\item{section}{Character; which section to find the number for.}
}
\value{
The number of the specified section.
}
\description{
Wiki sections must be specified by number for editing. Find the number, given
the name.
}
\examples{
get_section_number(word = "Wiktionary:Sandbox")
}
