% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_wiki.R
\name{get_section_text}
\alias{get_section_text}
\title{Get the Section Text for a Wiktionary Page}
\usage{
get_section_text(word, section_number)
}
\arguments{
\item{word}{Character; the Wiktionary page in question.}

\item{section_number}{Character; the number of the section to retrieve.}
}
\value{
The text of the specified section.
}
\description{
Given a page and the number of a section, return the text of the section.
}
\examples{
word <- "Wiktionary:Sandbox"
sn <- get_section_number(word)
get_section_text(word, sn)
}
