% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_wiki.R
\name{submit_morphology_edit}
\alias{submit_morphology_edit}
\title{Submit a Morphology Edit to a Wiktionary Page}
\usage{
submit_morphology_edit(
  word = "Project:Sandbox",
  template_text = "",
  username,
  password
)
}
\arguments{
\item{word}{Character; the Wiktionary page in question.}

\item{template_text}{Character; text to append to Etymology section.}

\item{username}{Character; bot username.}

\item{password}{Character; bot password.}
}
\value{
Logical; TRUE if the result of the edit POST request was "Success",
FALSE otherwise.
}
\description{
Given a Wiktionary page by title, and some text (generated by one of the
morphology templates), along with bot credentials, edits the Wiktionary
entry by appending the text at the end of the (English) Etymology section.
}
\examples{
\dontrun{
template_text = generate_morphology_suffix_template("converse", "ation")
submit_morphology_edit(word = "conversation",
                       template_text = template_text,
                       "usernamehere", "passwordhere")
}
}
